requirements.txt
# local package
-e .

click
Sphinx
coverage
flake8
python-dotenv>=0.5.1


matplotlib
setuptools

hydra-core
hydra_colorlog

seaborn
torch
pyro-ppl
pytorch-lightning
mlflow
numpy
scikit-learn

omegaconf
tqdm
ray[tune]
econml
xbcausalforest